Support --with-{zfs|ldiskfs} rpmbuild options
When rebuilding the source rpm package (or tarball) support
conditionally building --with and --without ldiskfs and zfs
support.  For compatibility with the existing automated test
infrastructure ldiskfs is enabled by default and zfs disabled.
This change also adds the kernel-devel package as a build
dependency.

Here's how to enable zfs support when rebuilding the source rpm.
$ rpmbuild --rebuild --with zfs lustre-x.y.z.src.rpm

Further documentation on rpm conditional builds is available here,
http://www.rpm.org/api/4.4.2.2/conditionalbuilds.html

Change-Id: I5481cc4a989512beb817173f5e1e68ae5607df38
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>