Handle undefined 'build_lustre_tests' in spec file
When 'build_lustre_tests' is undefined a parseExpressionBoolean
error occurs.  None of the checks which rely on 'build_lustre_tests'
have been written to handle to undefined case.  Presumably this
was overlooked because the automated build process always sets
it to 0 or 1 with a --define.  Update the code to assume undefined
means disabled.

  $ rpmbuild -ts lustre-2.1.0.tar.gz
  error: parse error in expression
  error: /.../lustre.spec:180: parseExpressionBoolean returns -1

Change-Id: I93aad468a674b7223407a5f7cc6ebf160e19e562
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>