ORI-375 tunefs: Call parse_opts after osd_read_ldd
The parse_opts function must be called after the osd_read_ldd function
is called. This is to ensure that the values read from disk are
overwritten by the values passes as options to the command. Otherwise,
the on disk options would always be used, regardless of what is passed
in to the command.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: I2f4c236849b82b6e112cc2f558e46ba6b92e6eb9