LU-8068 osd-zfs: large dnode support
In ZFS 0.7.x support for large dnode has been added.  This allows
Lustre to optionally specify the size of a dnode which ZFS will
then use to store metadata such as xattrs.

The default dnode size specified by the 'dnodesize' property on the
dataset will be used unless a specific value is provided.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>